+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 - Click [here](https://r35007.github.io/Mock-Server/) for mock server documentation.
 - Click [here](https://github.com/R35007/Mock-Server/blob/main/CHANGELOG.md) to see more changes reg the `@r35007/mock-server` package
 
+## v8.1.0
+
+- updated - `@r35007/mock-server@8.0.0`.
+- added - `mock-server.settings.homePage` - If false, The Mock Server Home Page will not be served. Default: true.
+
 ## v8.0.1
 
 - code optimized.